divide serialization
